+++ src/nespresso/__main__.py
@@ -10,6 +10,7 @@ from nespresso.bot.handlers.common.register import (
 from nespresso.bot.lib.notifications import admin
 from nespresso.bot.lib.notifications.erroring import SetExceptionHandlers
 from nespresso.bot.lib.notifications.pending import ProcessPendingUpdates
+from nespresso.bot.menu import SetMenu
 from nespresso.core.configs.paths import EnsurePaths
 from nespresso.core.logs import bot as logs
 from nespresso.db.session import InitDB
@@ -17,6 +18,7 @@ from nespresso.db.session import InitDB
 
 async def OnStartup() -> None:
     await InitDB()
+    await SetMenu()
     EnsurePaths()
 
     await logs.LoggerStart()