+++ src/nespresso/bot/menu.py
@@ -1,0 +1,13 @@
+from aiogram.types import BotCommand
+
+from nespresso.bot.creator import bot
+
+
+async def SetMenu() -> None:
+    commands = [
+        BotCommand(command="/start", description="???"),
+        BotCommand(command="/cancel", description="Cancel current state"),
+        BotCommand(command="/help", description="Help"),
+    ]
+
+    await bot.set_my_commands(commands)