+++ src/nespresso/services/user.py
@@ -1,4 +1,3 @@
-from nespresso.db.models.nes_user import NesUser
 from nespresso.db.models.tg_user import TgUser
 from nespresso.db.repositories.nes_user import NesUserRepository
 from nespresso.db.repositories.tg_user import TgUserRepository
@@ -12,12 +11,17 @@ class UserService:
         self.nes_user_repo = nes_user_repo
 
         # ----- Create -----
+        # - Tg -
+        self.RegisterTgUser = self.tg_user_repo.CreateTgUser
+
+        # - Nes -
+        self.UpsertNesUser = self.nes_user_repo.UpsertNesUsers
 
         # ----- Read -----
         # - Tg -
-        self.GetTgChatIdBy = self.tg_user_repo.GetChatIdBy
+        self.GetTgUsersOnCondition = self.tg_user_repo.GetTgUsersOnCondition
         self.GetTgUser = self.tg_user_repo.GetTgUser
-        self.GetTgUserColumn = self.tg_user_repo.GetTgUserColumn
+        self.GetTgChatIdBy = self.tg_user_repo.GetChatIdBy
 
         # - Nes -
         self.GetNesUser = self.nes_user_repo.GetNesUser
@@ -30,31 +34,32 @@ class UserService:
         # ----- Delete -----
 
     # ----- Create -----
-    # - Tg -
-    async def RegisterTgUser(
-        self, chat_id: int, username: str | None = None, full_name: str | None = None
-    ) -> None:
-        await self.tg_user_repo.CreateTgUser(
-            chat_id=chat_id, username=username, full_name=full_name
-        )
-
-    # - Nes -
-    async def UpsertNesUser(self, user: NesUser) -> None:
-        await self.nes_user_repo.UpsertNesUsers(users=user)
 
     # ----- Read -----
     # - Tg -
     async def CheckTgUserExists(self, chat_id: int) -> bool:
-        result = await self.tg_user_repo.GetTgUserColumn(chat_id, TgUser.chat_id)
+        result = await self.GetTgUser(
+            chat_id=chat_id,
+            column=TgUser.chat_id,
+        )
 
         return result is not None
 
     async def GetTgUsername(self, chat_id: int) -> str:
-        result = await self.tg_user_repo.GetTgUserColumn(
-            chat_id=chat_id, column=TgUser.username
+        result = await self.GetTgUser(
+            chat_id=chat_id,
+            column=TgUser.username,
         )
 
         return str(result)
+
+    async def GetRegisteredTgUsersChatId(self) -> list[int]:
+        result = await self.GetTgUsersOnCondition(
+            condition=TgUser.active_profile,
+            column=TgUser.chat_id,
+        )
+
+        return result if result else []
 
     # ----- Update -----