+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.0] - 2025-05-25 - Vlad
+### Updated
+- Functionality of TgUser repository
+
 ## [0.1.0] - 2025-05-24 - Vlad
 ### Added
 - Admin's command to send messages